split out tests into new modules
