+++ custom/opm/opm_reports/tests/__init__.py
@@ -1,2 +1,4 @@
 from .case_reports import *
+from .test_status import *
+from .test_windows_and_months import *
 # from .old_regression_tests import TestIncentive
